+++ src/inspector/components/Inspector/InspectorContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToPropsFunction } from "react-redux";
 import { IInspectorDispatch, IInspectorProps, Inspector } from "./Inspector";
 import { setMainTabAction, setModeTabAction } from "../../actions/inspectorActions";
 import { getMainTabID, getModeTabID, getActiveDescriptorContent, getActiveDescriptorReference } from "../../selectors/inspectorSelectors";
-import { IRootState } from "../../../store";
+import { IRootState } from "../../../shared/store";
 
 
 const mapStateToProps = (state: IRootState): IInspectorProps => {