+++ src/shared/store.tsx
@@ -1,9 +1,9 @@
 import { createStore, applyMiddleware, combineReducers } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
-import { inspectorReducer } from "./inspector/reducers";
-import { listenerReducer } from "./listener/reducers";
-import { IAppState } from "./listener/reducers/initialStateListener";
-import { IInspectorState } from "./inspector/model/types";
+import { inspectorReducer } from "../inspector/reducers";
+import { listenerReducer } from "../listener/reducers";
+import { IAppState } from "../listener/reducers/initialStateListener";
+import { IInspectorState } from "../inspector/model/types";
 
 /*
  * We're giving State interface to create store